refactor: moved file/zip save and export stuff to export_manager.js
