+++ src/cms/js/cms.js
@@ -46,6 +46,7 @@ module.exports = {
     this.sectionManager = require('modules/section_manager');
     this.metaManager    = require('modules/meta_manager');
     this.previewManager = require('modules/preview_manager');
+    this.exportManager  = require('modules/export_manager');
     this.templater      = require('modules/templater');
     this.vocabEditor    = require('modules/vocab_editor');
     this.ui             = require('modules/ui');
@@ -53,6 +54,7 @@ module.exports = {
     this.modal.init();
     this.vocabEditor.init();
     this.previewManager.init();
+    this.exportManager.init();
     if (!cms.showTranslation()){
       this.editor.init();
       this.videoManager.init();
@@ -110,88 +112,6 @@ module.exports = {
   showTranslation: function (){
     if (this.getQueryVariable('preview') != '') return true;
     return false;    
-  },
-
-  savePage: function(){
-    this.ui.hideMenu();
-    var html = cms.getPageHTMLWithoutCMS();
-    html = cms.addDocType(html);
-    cms.saveHtmlToFile(html, cms.saveToZip);
-  },
-
-  getPageHTMLWithoutCMS: function () {
-    var cleanHTML = '',
-        $html = $('html').clone();
-
-    cms.config.editableItems.forEach(function (el) {
-      $html.find(el+':empty').remove();
-    });
-    // remove elems added by cms
-    $html.find('.cms-menu-container, .cms-menu, .cms-menu-bg, .cms-modal, .cms-media-btn, .cms-menu-btn').remove();
-    // remove all classes and attributes
-    $html.find('*').removeClass('cms-editable cms-editable-img cms-editable-region cms-inline-media');
-    $html.find('*').removeClass(cms.config.mustardClass);
-    $html.find('*').removeAttr('contenteditable');
-    $html.find('*').removeAttr('spellcheck');
-    // remove cms scripts
-    $html.find('script[src^="cms"], #cms-init, link[href^="cms"]').remove();
-    $html.find('*[class=""]').removeAttr('class');
-    // reset app templates so they work on pages with no js
-    // move to a method in the main app
-    $html.find('body').removeClass('js');
-    $html.find('*').removeClass('anim-fade-1s transparent scrollmation-text-js scrollmation-image-container-top scrollmation-image-container-fixed scrollmation-image-container-bottom');
-    $html.find('.scrollmation-text').addClass('article');
-    $html.find('.video-overlay').removeClass('hidden');
-    $html.find('.video-overlay-button').html('▶');
-    // get cleaned html
-    cleanHTML = $html.html();
-    cleanHTML = cms.removeWhitespace(cleanHTML);
-
-    return cleanHTML;
-  },
-
-  addDocType: function (html) {
-    var lang = cms.vocabEditor.getCurrentService() || 'en';
-    return '<!DOCTYPE html>\n<html lang="'+lang+'">\n' + html + '</html>';
-  },
-
-  removeWhitespace: function(string){
-    string = string.replace(/  /g, '');
-    string = string.replace(/ \n/g, '\n');
-    string = string.replace(/\n\n/g, '');
-    return string;
-  },
-
-  saveHtmlToFile: function(html, callback) {
-    var data = new FormData();
-    data.append('html', html);
-
-    this.ajax.create('POST', 'cms/api/preview.php');
-    var successHandler = function (responseText) {
-      console.log(responseText);
-      callback();
-    }
-    var errorHandler = function (responseText) {
-      console.log(responseText);
-    }
-    this.ajax.onFinish(successHandler, errorHandler);
-    this.ajax.send(data);
-  },
-
-  saveToZip: function () {
-    var data = new FormData();
-    data.append('savetozip', 'true');
-
-    cms.ajax.create('POST', 'cms/api/save.php');
-    var successHandler = function (responseText) {
-      console.log(responseText);
-      window.location = responseText;
-    }
-    var errorHandler = function (responseText) {
-      console.log(responseText);
-    }
-    cms.ajax.onFinish(successHandler, errorHandler);
-    cms.ajax.send(data);
   },
 
   autoSave: function () {
